test: merged authorship test in rewards.test.ts
