+++ tests/rewards.test.ts
@@ -10,6 +10,7 @@ import { ContextPlugin } from "../src/types/plugin-input";
 import { db } from "./__mocks__/db";
 import dbSeed from "./__mocks__/db-seed.json";
 import { server } from "./__mocks__/node";
+import authorshipRewardResult from "./__mocks__/results/authorship-reward.json";
 import rewardSplitResult from "./__mocks__/results/reward-split.json";
 import cfg from "./__mocks__/results/valid-configuration.json";
 import "./helpers/permit-mock";
@@ -58,30 +59,32 @@ jest.unstable_mockModule("@supabase/supabase-js", () => {
   };
 });
 
+const collectLinkedMergedPulls = jest.fn(() => [
+  {
+    id: "PR_kwDOKzVPS85zXUoj",
+    title: "fix: add state to sorting manager for bottom and top",
+    number: 70,
+    url: "https://github.com/ubiquity/work.ubq.fi/pull/70",
+    author: {
+      login: "0x4007",
+      id: 4975670,
+    },
+    state: "MERGED",
+    repository: {
+      owner: {
+        login: "ubiquity",
+      },
+      name: "work.ubq.fi",
+    },
+  },
+]);
+
 jest.unstable_mockModule("../src/helpers/get-comment-details", () => ({
   getMinimizedCommentStatus: jest.fn(),
 }));
 
 jest.unstable_mockModule("../src/data-collection/collect-linked-pulls", () => ({
-  collectLinkedMergedPulls: jest.fn(() => [
-    {
-      id: "PR_kwDOKzVPS85zXUoj",
-      title: "fix: add state to sorting manager for bottom and top",
-      number: 70,
-      url: "https://github.com/ubiquity/work.ubq.fi/pull/70",
-      author: {
-        login: "0x4007",
-        id: 4975670,
-      },
-      state: "MERGED",
-      repository: {
-        owner: {
-          login: "ubiquity",
-        },
-        name: "work.ubq.fi",
-      },
-    },
-  ]),
+  collectLinkedMergedPulls: collectLinkedMergedPulls,
 }));
 
 beforeAll(() => {
@@ -190,6 +193,45 @@ describe("Rewards tests", () => {
     expect(result).toEqual(rewardSplitResult);
     expect(fs.readFileSync("./output.html", "utf-8")).toEqual(
       fs.readFileSync("./tests/__mocks__/results/output-reward-split.html", "utf-8")
+    );
+  }, 120000);
+
+  it("Should distribute rewards based on authorship percentages in issue body edits", async () => {
+    collectLinkedMergedPulls.mockReturnValueOnce([
+      {
+        id: "PR_kwDOKzVPS85zXUoj",
+        title: "fix: add authorship to issue body edits",
+        number: 101,
+        url: "https://github.com/ubiquity/work.ubq.fi/pull/101",
+        author: {
+          login: "contributor1",
+          id: 4975670,
+        },
+        state: "MERGED",
+        repository: {
+          owner: {
+            login: "ubiquity",
+          },
+          name: "work.ubq.fi",
+        },
+      },
+    ]);
+    const issue = parseGitHubUrl("https://github.com/ubiquity/work.ubq.fi/issues/100");
+    const activity = new IssueActivity(ctx, issue);
+    await activity.init();
+    const processor = new Processor(ctx);
+    processor["_transformers"] = [
+      new UserExtractorModule(ctx),
+      new DataPurgeModule(ctx),
+      new FormattingEvaluatorModule(ctx),
+      new ContentEvaluatorModule(ctx),
+      new GithubCommentModule(ctx),
+    ];
+    await processor.run(activity);
+    const result = JSON.parse(processor.dump());
+    expect(result).toEqual(authorshipRewardResult);
+    expect(fs.readFileSync("./output.html", "utf-8")).toEqual(
+      fs.readFileSync("./tests/__mocks__/results/output-authorship-reward.html", "utf-8")
     );
   }, 120000);
 });
